[Form] Fixed: Added "validation_groups" option to submit button
